Refactor entire library codebase for smallers mods
