+++ test/ex_check/project_cases/missing_test_helper_test.exs
@@ -13,7 +13,7 @@ defmodule ExCheck.ProjectCases.MissingTestHelperTest do
 
     assert String.contains?(output, "compiler success")
     assert String.contains?(output, "formatter success")
-    assert String.contains?(output, "ex_unit skipped due to missing file test/test_helper.exs")
+    assert String.contains?(output, "ex_unit skipped due to missing file ./test/test_helper.exs")
     assert String.contains?(output, "credo skipped due to missing package credo")
     assert String.contains?(output, "sobelow skipped due to missing package sobelow")
     assert String.contains?(output, "dialyzer skipped due to missing package dialyxir")
